+++ cli.py
@@ -2,7 +2,6 @@ from sqlalchemy import create_engine
 from sqlalchemy.orm import sessionmaker, scoped_session
 
 from hydrus.app_factory import app_factory
-from hydrus.socketio_factory import create_socket
 from hydrus.utils import (set_session, set_doc, set_hydrus_server_url,
                           set_token, set_api_name, set_authentication,
                           set_page_size, set_pagination)
@@ -11,7 +10,6 @@ from hydra_python_core import doc_maker
 from hydrus.data.db_models import Base
 from hydrus.data.user import add_user
 from hydrus.data.exceptions import UserExists
-from hydrus.data.stale_records_cleanup import remove_stale_modification_records
 from gevent.pywsgi import WSGIServer
 from hydra_openapi_parser.openapi_parser import parse
 from hydrus.samples.hydra_doc_sample import doc as api_document
@@ -29,7 +27,7 @@ import yaml
               help="The API name.", type=str)
 @click.option("--auth/--no-auth", default=True,
               help="Set authentication to True or False.")
-@click.option("--dburl", default="sqlite:///database.db",
+@click.option("--dburl", default="sqlite:///:memory:",
               help="Set database url", type=str)
 @click.option("--hydradoc", "-d", default=None,
               help="Location to HydraDocumentation (JSON-LD) of server.",
@@ -44,13 +42,10 @@ import yaml
               help="Toggle token based user authentication.")
 @click.option("--serverurl", default="http://localhost",
               help="Set server url", type=str)
-@click.option("--stale_records_removal_interval", default=900,
-              help="Interval period between removal of stale modification records.",
-              type=int)
 @click.argument("serve", required=True)
 def startserver(adduser: Tuple, api: str, auth: bool, dburl: str, pagination: bool,
                 hydradoc: str, port: int, pagesize: int, serverurl: str, token: bool,
-                stale_records_removal_interval: int, serve: None) -> None:
+                serve: None) -> None:
     """
     Python Hydrus CLI
 
@@ -87,8 +82,8 @@ def startserver(adduser: Tuple, api: str, auth: bool, dburl: str, pagination: bo
 
     click.echo("Setting up the database")
     # Create a connection to the database you want to use
-    engine = create_engine(DB_URL, connect_args={'check_same_thread': False})
-    Base.metadata.drop_all(engine)
+    engine = create_engine(DB_URL)
+
     click.echo("Creating models")
     # Add the required Models to the database
     Base.metadata.create_all(engine)
@@ -164,8 +159,6 @@ def startserver(adduser: Tuple, api: str, auth: bool, dburl: str, pagination: bo
     # Create a Hydrus app with the API name you want, default will be "api"
     app = app_factory(API_NAME)
     # Set the name of the API
-    # Create a socket for the app
-    socketio = create_socket(app, session)
     click.echo("Starting the application")
     with set_authentication(app, auth):
         # Use authentication for all requests
@@ -181,17 +174,17 @@ def startserver(adduser: Tuple, api: str, auth: bool, dburl: str, pagination: bo
                             with set_pagination(app, pagination):
                                 # Set page size of a collection view
                                 with set_page_size(app, pagesize):
-                                    # Run a thread to remove stale modification records at some
-                                    # interval of time.
-                                    remove_stale_modification_records(
-                                        session, stale_records_removal_interval)
                                     # Start the hydrus app
-                                    socketio.run(app, port=port)
+                                    http_server = WSGIServer(('', port), app)
                                     click.echo("Server running at:")
                                     click.echo(
                                         "{}{}".format(
                                             HYDRUS_SERVER_URL,
                                             API_NAME))
+                                    try:
+                                        http_server.serve_forever()
+                                    except KeyboardInterrupt:
+                                        pass
 
 
 if __name__ == "__main__":
