+++ hydrus/data/crud.py
@@ -31,12 +31,12 @@
     typing : Module which provides support for type hints .
 
 """  # nopep8
-import re
+
 from sqlalchemy.orm import with_polymorphic
 from sqlalchemy import exists
 from sqlalchemy.orm.exc import NoResultFound
 from hydrus.data.db_models import (Graph, BaseProperty, RDFClass, Instance,
-                                   Terminal, GraphIAC, GraphIIT, GraphIII, Modification)
+                                   Terminal, GraphIAC, GraphIIT, GraphIII)
 from hydrus.data.exceptions import (
     ClassNotFound,
     InstanceExists,
@@ -48,15 +48,9 @@ from hydrus.data.exceptions import (
     InvalidSearchParameter,
     IncompatibleParameters,
     OffsetOutOfRange)
-from hydrus.data.crud_helpers import (
-    apply_filter,
-    recreate_iri,
-    attach_hydra_view,
-    pre_process_pagination_parameters,
-    parse_search_params)
 # from sqlalchemy.orm.session import Session
 from sqlalchemy.orm.scoping import scoped_session
-from typing import Dict, Optional, Any, List
+from typing import Dict, Optional, Any, List, Tuple
 
 triples = with_polymorphic(Graph, '*')
 properties = with_polymorphic(BaseProperty, "*")
@@ -138,11 +132,10 @@ def get(id_: str, type_: str, api_name: str, session: scoped_session,
     return object_template
 
 
-def insert(object_: Dict[str, Any], session: scoped_session, link_props: Dict[str, Any]={},
+def insert(object_: Dict[str, Any], session: scoped_session,
            id_: Optional[str] = None) -> str:
     """Insert an object to database [POST] and returns the inserted object.
     :param object_: object to be inserted
-    :param link_props: Hydra link properties in the object.
     :param session: sqlalchemy scoped session
     :param id_: id of the object to be inserted (optional param)
     :return: ID of object inserted
@@ -186,37 +179,59 @@ def insert(object_: Dict[str, Any], session: scoped_session, link_props: Dict[st
                 # Adds new Property
                 session.close()
                 raise PropertyNotFound(type_=prop_name)
-            # For insertion in III through link
-            if prop_name in link_props:
-                try:
-                    insert_iii_with_link(instance.id, property_,
-                                         link_props[prop_name], session)
-                except (NotInstanceProperty, InstanceNotFound, ClassNotFound):
-                    raise
-                continue
+
             # For insertion in III
             if isinstance(object_[prop_name], dict):
-                try:
-                    insert_iii(object_=object_, prop_name=prop_name, instance=instance,
-                               property_=property_, session=session)
-                except NotInstanceProperty:
-                    raise
+                instance_id = insert(object_[prop_name], session=session)
+                instance_object = session.query(Instance).filter(
+                    Instance.id == instance_id).one()
+                if property_.type_ == "PROPERTY" or property_.type_ == "INSTANCE":
+                    property_.type_ = "INSTANCE"
+                    session.add(property_)
+                    triple = GraphIII(
+                        subject=instance.id,
+                        predicate=property_.id,
+                        object_=instance_object.id)
+                    session.add(triple)
+                else:
+                    session.close()
+                    raise NotInstanceProperty(type_=prop_name)
+
             # For insertion in IAC
             elif session.query(exists().where(RDFClass.name == str(object_[prop_name]))).scalar() \
                     and property_.type_ == "PROPERTY" or property_.type_ == "ABSTRACT":
-                try:
-                    insert_iac(object_=object_, prop_name=prop_name, instance=instance,
-                               property_=property_, session=session)
-                except NotAbstractProperty:
-                    raise
+                property_.type_ = "ABSTRACT"
+                session.add(property_)
+                class_ = session.query(RDFClass).filter(
+                    RDFClass.name == object_[prop_name]).one()
+                triple = GraphIAC(
+                    subject=instance.id,
+                    predicate=property_.id,
+                    object_=class_.id)
+                session.add(triple)
+            elif session.query(exists().where(RDFClass.name == str(object_[prop_name]))).scalar():
+                session.close()
+                raise NotAbstractProperty(type_=prop_name)
 
             # For insertion in IIT
             else:
-                try:
-                    insert_iit(object_=object_, prop_name=prop_name, instance=instance,
-                               property_=property_, session=session)
-                except NotInstanceProperty:
-                    raise
+                terminal = Terminal(value=object_[prop_name])
+                session.add(terminal)
+                session.flush()  # Assigns ID without committing
+
+                if property_.type_ == "PROPERTY" or property_.type_ == "INSTANCE":
+                    property_.type_ = "INSTANCE"
+                    session.add(property_)
+                    triple = GraphIIT(
+                        subject=instance.id,
+                        predicate=property_.id,
+                        object_=terminal.id)
+                    # Add things directly to session, if anything fails whole
+                    # transaction is aborted
+                    session.add(triple)
+                else:
+                    session.close()
+                    raise NotInstanceProperty(type_=prop_name)
 
     session.commit()
     return instance.id
@@ -225,13 +240,11 @@ def insert(object_: Dict[str, Any], session: scoped_session, link_props: Dict[st
 def insert_multiple(objects_: List[Dict[str,
                                         Any]],
                     session: scoped_session,
-                    link_props_list: List[Dict[str, Any]]=[],
                     id_: Optional[str] = "") -> List[str]:
     """
     Adds a list of object with given ids to the database
     :param objects_: List of dict's to be added to the database
     :param session: scoped session from getSession in utils
-    :param link_props_list: List of link properties for each object being inserted.
     :param id_: optional parameter containing the ids of objects that have to be inserted
     :return: Ids that have been inserted
 
@@ -293,52 +306,63 @@ def insert_multiple(objects_: List[Dict[str,
                     # Adds new Property
                     session.close()
                     raise PropertyNotFound(type_=prop_name)
-                if len(link_props_list) > 0:
-                    # For insertion in III through link
-                    if prop_name in link_props_list[index]:
-                        try:
-                            triple = insert_iii_with_link(instances[index].id,
-                                                          property_,
-                                                          link_props_list[index][prop_name],
-                                                          session)
-                            triples_list.append(triple)
-                            properties_list.append(property_)
-                        except (NotInstanceProperty, InstanceNotFound, ClassNotFound):
-                            raise
-                        continue
+
                 # For insertion in III
                 if isinstance(objects_[index][prop_name], dict):
-                    try:
-                        triple = insert_iii(object_=objects_[index], prop_name=prop_name,
-                                            instance=instances[index], property_=property_,
-                                            session=session)
-                        triples_list.append(triple)
+                    instance_id = insert(
+                        objects_[index][prop_name], session=session)
+                    instance_object = session.query(Instance).filter(
+                        Instance.id == instance_id).one()
+
+                    if property_.type_ == "PROPERTY" or property_.type_ == "INSTANCE":
+                        property_.type_ = "INSTANCE"
                         properties_list.append(property_)
-                    except NotInstanceProperty:
-                        raise
+                        triple = GraphIII(
+                            subject=instances[index].id,
+                            predicate=property_.id,
+                            object_=instance_object.id)
+                        triples_list.append(triple)
+                    else:
+                        session.close()
+                        raise NotInstanceProperty(type_=prop_name)
 
                 # For insertion in IAC
                 elif session.query(
                         exists().where(RDFClass.name == str(objects_[index][prop_name]))).scalar():
-                    try:
-                        triple = insert_iac(object_=objects_[index], prop_name=prop_name,
-                                            instance=instances[index], property_=property_,
-                                            session=session)
-                        triples_list.append(triple)
+                    if property_.type_ == "PROPERTY" or property_.type_ == "ABSTRACT":
+                        property_.type_ = "ABSTRACT"
                         properties_list.append(property_)
-                    except NotAbstractProperty:
-                        raise
+                        class_ = session.query(RDFClass).filter(
+                            RDFClass.name == objects_[index][prop_name]).one()
+                        triple = GraphIAC(
+                            subject=instances[index].id,
+                            predicate=property_.id,
+                            object_=class_.id)
+                        triples_list.append(triple)
+
+                    else:
+                        session.close()
+                        raise NotAbstractProperty(type_=prop_name)
 
                 # For insertion in IIT
                 else:
-                    try:
-                        triple = insert_iit(object_=objects_[index], prop_name=prop_name,
-                                            instance=instances[index], property_=property_,
-                                            session=session)
-                        triples_list.append(triple)
+                    terminal = Terminal(value=objects_[index][prop_name])
+                    session.add(terminal)
+                    session.flush()  # Assigns ID without committing
+
+                    if property_.type_ == "PROPERTY" or property_.type_ == "INSTANCE":
+                        property_.type_ = "INSTANCE"
                         properties_list.append(property_)
-                    except NotInstanceProperty:
-                        raise
+                        triple = GraphIIT(
+                            subject=instances[index].id,
+                            predicate=property_.id,
+                            object_=terminal.id)
+                        # Add things directly to session, if anything fails
+                        # whole transaction is aborted
+                        triples_list.append(triple)
+                    else:
+                        session.close()
+                        raise NotInstanceProperty(type_=prop_name)
     session.bulk_save_objects(properties_list)
     session.bulk_save_objects(triples_list)
     session.commit()
@@ -464,7 +488,6 @@ def update(id_: str,
                          str],
            session: scoped_session,
            api_name: str,
-           link_props: Dict[str, Any]={},
            path: str = None) -> str:
     """Update an object properties based on the given object [PUT].
     :param id_: if of object to be updated
@@ -472,7 +495,6 @@ def update(id_: str,
     :param object_: object that has to be inserted
     :param session: sqlalchemy scoped session
     :param api_name: api name specified while starting server
-    :param link_props: Link properties of the object being updated.
     :param path: endpoint
     :return: id of updated object
     """
@@ -483,10 +505,10 @@ def update(id_: str,
     delete(id_=id_, type_=type_, session=session)
     # Try inserting new object
     try:
-        insert(object_=object_, id_=id_, link_props=link_props, session=session)
+        insert(object_=object_, id_=id_, session=session)
     except (ClassNotFound, InstanceExists, PropertyNotFound) as e:
         # Put old object back
-        insert(object_=instance, id_=id_, link_props=link_props, session=session)
+        insert(object_=instance, id_=id_, session=session)
         raise e
 
     get(id_=id_, type_=type_, session=session, api_name=api_name, path=path)
@@ -516,8 +538,7 @@ def get_collection(API_NAME: str,
     """
     try:
         # Reconstruct dict with property ids as keys
-        search_props = parse_search_params(search_params=search_params, properties=properties,
-                                           session=session)
+        search_props = parse_search_params(search_params=search_params, session=session)
     except InvalidSearchParameter:
         raise
 
@@ -541,7 +562,7 @@ def get_collection(API_NAME: str,
         instances = list()
     filtered_instances = list()
     for instance_ in instances:
-        if apply_filter(instance_.id, search_props, triples, session) is True:
+        if apply_filter(instance_.id, session, search_props=search_props) is True:
             filtered_instances.append(instance_)
     result_length = len(filtered_instances)
     try:
@@ -658,13 +679,11 @@ def update_single(object_: Dict[str,
                                 Any],
                   session: scoped_session,
                   api_name: str,
-                  link_props: Dict[str, Any],
                   path: str = None) -> int:
     """Update instance of classes with single objects.
     :param object_: new object
     :param session: sqlalchemy scoped session
     :param api_name: api name specified while starting server
-    :param link_props: Link properties of the object being updated
     :param path: endpoint
     :return: id of the updated object
 
@@ -691,7 +710,6 @@ def update_single(object_: Dict[str,
         object_=object_,
         session=session,
         api_name=api_name,
-        link_props=link_props,
         path=path)
 
 
@@ -721,215 +739,194 @@ def delete_single(type_: str, session: scoped_session) -> None:
     return delete(instance.id, type_, session=session)
 
 
-def insert_modification_record(method: str, resource_url: str,
-                               session: scoped_session) -> int:
+def apply_filter(object_id: str, session: scoped_session,
+                 search_props: Dict[str, Any]) -> bool:
+    """Check whether objects has properties with query values or not.
+    :param object_id: Id of the instance.
+    :param session: sqlalchemy scoped session.
+    :param search_props: Dictionary of query parameters with property id and values.
+    :return: True if the instance has properties with given values, False otherwise.
     """
-    Insert a modification record into the database.
-    :param method: HTTP method type of related operation.
-    :param resource_url: URL of resource modified.
+    for prop in search_props:
+        # For nested properties
+        if isinstance(search_props[prop], dict):
+            data = session.query(triples).filter(
+                triples.GraphIII.subject == object_id, triples.GraphIII.predicate == prop).one()
+            if apply_filter(data.object_, session, search_props=search_props[prop]) is False:
+                return False
+        else:
+            data = session.query(triples).filter(
+                triples.GraphIIT.subject == object_id, triples.GraphIIT.predicate == prop).one()
+            terminal = session.query(Terminal).filter(
+                Terminal.id == data.object_).one()
+            if terminal.value != search_props[prop]:
+                return False
+    return True
+
+
+def recreate_iri(API_NAME: str, path: str, search_params: Dict[str, Any]) -> str:
+    """Recreate the IRI with query arguments
+    :param API_NAME: API name specified while starting the server.
+    :param path: endpoint
+    :param search_params: List of query parameters.
+    :return: Recreated IRI.
+    """
+    iri = "/{}/{}?".format(API_NAME, path)
+    for param in search_params:
+        # Skip page, pageIndex or offset parameters as they will be updated to point to
+        # next, previous and last page
+        if param == "page" or param == "pageIndex" or param == "offset":
+            continue
+        iri += "{}={}&".format(param, search_params[param])
+    return iri
+
+
+def parse_search_params(search_params: Dict[str, Any],
+                        session: scoped_session) -> Dict[str, Any]:
+    """Parse search parameters and create a dict with id of parameters as keys.
+    :param search_params: Dictionary having input search parameters.
     :param session: sqlalchemy session.
-    :return: ID of new modification record.
+    :return: A dictionary having property ids as keys.
     """
-    modification = Modification(method=method, resource_url=resource_url)
-    session.add(modification)
-    session.commit()
-    return modification.job_id
-
-
-def get_last_modification_job_id(session: scoped_session) -> str:
-    """
-    Get job id of most recent modification record stored in the db.
-    :param session: sqlalchemy session
-    :return: job id of recent modification.
-    """
-    last_modification = session.query(Modification).order_by(Modification.job_id.desc()).first()
-    if last_modification is None:
-        last_job_id = ""
-    else:
-        last_job_id = last_modification.job_id
-    return last_job_id
-
-
-def get_modification_table_diff(session: scoped_session,
-                                agent_job_id: str = None) -> List[Dict[str, Any]]:
-    """
-    Get modification table difference.
-    :param session: sqlalchemy session.
-    :param agent_job_id: Job id from the client.
-    :return: List of all modifications done after job with job_id = agent_job_id.
-    """
-    # If agent_job_id is not given then return all the elements.
-    if agent_job_id is None:
-        modifications = session.query(Modification).order_by(
-            Modification.job_id.asc()).all()
-    # If agent_job_id is given then return all records which are older
-    # than the record with agent_job_id.
-    else:
-        try:
-            record_for_agent_job_id = session.query(Modification).filter(
-                Modification.job_id == agent_job_id).one()
-        except NoResultFound:
-            return []
-        modifications = session.query(Modification).filter(
-            Modification.job_id > record_for_agent_job_id.job_id).order_by(
-            Modification.job_id.asc()).all()
-
-    # Create response body
-    list_of_modification_records = []
-    for modification in modifications:
-        modification_record = {
-            "job_id": modification.job_id,
-            "method": modification.method,
-            "resource_url": modification.resource_url
-        }
-        list_of_modification_records.append(modification_record)
-    return list_of_modification_records
-
-
-def insert_iii(object_: Dict[str, Any], prop_name: str,
-               instance: Instance, property_: BaseProperty,
-               session: scoped_session) -> GraphIII:
-    """
-    Insert a GraphIII triple in the database.
-    :param object_:  Object body.
-    :param prop_name: Property name.
-    :param instance: Instance for the newly added object.
-    :param property_: Predicate in the new triple being inserted.
-    :param session: sqlalchemy session.
-
-    :return: GraphIII triple.
-
-    :raises: NotInstanceProperty
-    """
-    instance_id = insert(object_[prop_name], session=session)
-    instance_object = session.query(Instance).filter(
-        Instance.id == instance_id).one()
-    if property_.type_ == "PROPERTY" or property_.type_ == "INSTANCE":
-        property_.type_ = "INSTANCE"
-        session.add(property_)
-        triple = GraphIII(
-            subject=instance.id,
-            predicate=property_.id,
-            object_=instance_object.id)
-        session.add(triple)
-        return triple
-    else:
-        session.close()
-        raise NotInstanceProperty(type_=prop_name)
-
-
-def insert_iac(object_: Dict[str, Any], prop_name: str,
-               instance: Instance, property_: BaseProperty,
-               session: scoped_session) -> GraphIAC:
-    """
-    Insert a GraphIAC triple in the database.
-    :param object_:  Object body.
-    :param prop_name: Property name.
-    :param instance: Instance for the newly added object.
-    :param property_: Predicate in the new triple being inserted.
-    :param session: sqlalchemy session.
-
-    :return: GraphIAC triple.
-    """
-    if property_.type_ == "PROPERTY" or property_.type_ == "ABSTRACT":
-        property_.type_ = "ABSTRACT"
-        session.add(property_)
-        class_ = session.query(RDFClass).filter(
-            RDFClass.name == object_[prop_name]).one()
-        triple = GraphIAC(
-            subject=instance.id,
-            predicate=property_.id,
-            object_=class_.id)
-        session.add(triple)
-        return triple
-    else:
-        session.close()
-        raise NotAbstractProperty(type_=prop_name)
-
-
-def insert_iit(object_: Dict[str, Any], prop_name: str,
-               instance: Instance, property_: BaseProperty,
-               session: scoped_session) -> GraphIIT:
-    """
-    Insert a GraphIIT triple in the database.
-    :param object_:  Object body.
-    :param prop_name: Property name.
-    :param instance: Instance for the newly added object.
-    :param property_: Predicate in the new triple being inserted.
-    :param session: sqlalchemy session.
-
-    :return: GraphIIT triple.
-
-    :raises: NotInstanceProperty
-    """
-    terminal = Terminal(value=object_[prop_name])
-    session.add(terminal)
-    session.flush()  # Assigns ID without committing
-
-    if property_.type_ == "PROPERTY" or property_.type_ == "INSTANCE":
-        property_.type_ = "INSTANCE"
-        session.add(property_)
-        triple = GraphIIT(
-            subject=instance.id,
-            predicate=property_.id,
-            object_=terminal.id)
-        # Add things directly to session, if anything fails whole
-        # transaction is aborted
-        session.add(triple)
-        return triple
-    else:
-        session.close()
-        raise NotInstanceProperty(type_=prop_name)
-
-
-def insert_iii_with_link(instance_id: str, property_: BaseProperty,
-                         property_value: str, session: scoped_session):
-    """
-    Inserts GraphIII triple to store a relation defined with hydra:Link.
-    :param instance_id:  Id of the instance being inserted
-    :param property_: Property being used as predicate in the new triple.
-    :param property_value: Value of the property being inserted.
-    :param session: sqlalchemy session
-    :return:
-    """
-    if property_.type_ == "PROPERTY" or property_.type_ == "INSTANCE":
-        property_.type_ = "INSTANCE"
-        # If value matches with the regex then value is an id and link is to an
-        # instance of a collection class otherwise value is a class_type and link
-        # is to a non collection class.
-        regex = r'[a-z0-9]{8}-([a-z0-9]{4}-){3}[a-z0-9]{12}'
-        matchObj = re.match(regex, property_value)
-        # Link is to an instance of a collection class
-        if matchObj:
+    search_props = dict()
+    pagination_parameters = ["page", "pageIndex", "limit", "offset"]
+    for param in search_params:
+        # Skip if the parameter is a pagination parameter
+        if param in pagination_parameters:
+            continue
+        # For one level deep nested parameters
+        if "[" in param and "]" in param:
+            prop_name = param.split('[')[0]
             try:
-                nested_instance = session.query(Instance).filter(
-                    Instance.id == property_value).one()
+                prop_id = session.query(properties).filter(
+                    properties.name == prop_name).one().id
+                if prop_id not in search_props:
+                    search_props[prop_id] = {}
+                nested_prop_id = session.query(properties).filter(
+                    properties.name == param[param.find('[') + 1:param.find(']')]).one().id
+                search_props[prop_id][nested_prop_id] = search_params[param]
             except NoResultFound:
-                raise InstanceNotFound(id_=property_value, type_="")
-            triple = GraphIII(
-                subject=instance_id,
-                predicate=property_.id,
-                object_=nested_instance.id)
-            session.add(triple)
-            return triple
-        # Link is to a non collection, single instance class
+                raise InvalidSearchParameter(param)
+        # For normal parameters
         else:
             try:
-                nested_rdf_class = session.query(RDFClass).filter(
-                    RDFClass.name == property_value).one()
+                prop_id = session.query(properties).filter(
+                    properties.name == param).one().id
+                search_props[prop_id] = search_params[param]
             except NoResultFound:
-                raise ClassNotFound(type_=property_value)
-            try:
-                nested_instance = session.query(Instance).filter(
-                    Instance.type_ == nested_rdf_class.id).all()[-1]
-            except (NoResultFound, IndexError, ValueError):
-                raise InstanceNotFound(type_=nested_rdf_class.name)
-            triple = GraphIII(
-                subject=instance_id,
-                predicate=property_.id,
-                object_=nested_instance.id)
-            session.add(triple)
-            return triple
+                raise InvalidSearchParameter(param)
+    return search_props
+
+
+def calculate_page_limit_and_offset(paginate: bool, page_size: int, page: int, result_length: int,
+                                    offset: int, limit: int) -> Tuple[int, int]:
+    """Calculate page limit and offset for pagination.
+    :param paginate: Showing whether pagination is enable/disable.
+    :param page_size: Number maximum elements showed in a page.
+    :param page: page number.
+    :param result_length: Length of the list containing desired elements.
+    :param offset: offset value.
+    :param limit: page limit.
+    :return: page limit and offset.
+    """
+    if limit is not None:
+        page_size = limit
+    if paginate is True:
+        if offset is None:
+            offset = (page - 1) * page_size
+        limit = page_size
     else:
-        session.close()
-        raise NotInstanceProperty(type_=property_.name)
+        offset = 0
+        limit = result_length
+
+    return limit, offset
+
+
+def pre_process_pagination_parameters(search_params: Dict[str, Any], paginate: bool,
+                                      page_size: int, result_length: int) -> Tuple[int, int, int]:
+    """Pre-process pagination related query parameters passed by client.
+    :param search_params: Dict of all search parameters.
+    :param paginate: Indicates if pagination is enabled/disabled.
+    :param page_size: Maximum element a page can contain.
+    :param result_length: Length of the list of containing desired items.
+    :return: returns page number, page limit and offset.
+    """
+    incompatible_parameters = ["page", "pageIndex", "offset"]
+    incompatible_parameters_len = len(incompatible_parameters)
+    # Find any pair of incompatible parameters
+    for i in range(incompatible_parameters_len):
+        if incompatible_parameters[i] not in search_params:
+            continue
+        if i != incompatible_parameters_len - 1:
+            for j in range(i+1, incompatible_parameters_len):
+                if incompatible_parameters[j] in search_params:
+                    raise IncompatibleParameters([incompatible_parameters[i],
+                                                  incompatible_parameters[j]])
+    try:
+        # Extract page number from query arguments
+        if "pageIndex" in search_params:
+            page = int(search_params.get("pageIndex"))
+            offset = None
+        elif "offset" in search_params:
+            offset = int(search_params.get("offset"))
+            page = offset//page_size + 1
+            if offset > result_length:
+                raise OffsetOutOfRange(str(offset))
+        else:
+            page = int(search_params.get("page", 1))
+            offset = None
+        if "limit" in search_params:
+            limit = int(search_params.get("limit"))
+        else:
+            limit = None
+    except ValueError:
+        raise PageNotFound(page)
+    page_limit, offset = calculate_page_limit_and_offset(paginate=paginate, page_size=page_size,
+                                                         page=page, result_length=result_length,
+                                                         offset=offset, limit=limit)
+    return page, page_limit, offset
+
+
+def attach_hydra_view(collection_template: Dict[str, Any], paginate_param: str,
+                      result_length: int, page_size: int,
+                      iri: str, offset: int = None,
+                      page: int = None, last: int = None) -> None:
+    """Attaches hydra:view to the collection template.
+    :param collection_template: the collection template.
+    :param paginate_param: type of paginate parameter used.
+    :param result_length: length of the result set.
+    :param page_size: size of the page.
+    :param iri: IRI of the collection with query parameters except "page", "pageIndex" and "offset".
+    :param offset: offset used for pagination, None if not used.
+    :param page: page number used for pagination, None if not used.
+    :param last: Page number of the last page only used when "page" or "pageIndex"
+                 is used for pagination, None otherwise.
+    """
+    if paginate_param == "offset":
+        collection_template["view"] = {
+            "@id": "{}{}={}".format(iri, paginate_param, offset),
+            "@type": "PartialCollectionView",
+            "first": "{}{}=0".format(iri, paginate_param),
+            "last": "{}{}={}".format(iri, paginate_param, result_length-page_size)
+        }
+        if offset > page_size:
+            collection_template["view"]["previous"] = "{}{}={}".format(iri,
+                                                                       paginate_param,
+                                                                       offset - page_size)
+        if offset < result_length-page_size:
+            collection_template["view"]["next"] = "{}{}={}".format(iri,
+                                                                   paginate_param,
+                                                                   offset + page_size)
+    else:
+        collection_template["view"] = {
+            "@id": "{}{}={}".format(iri, paginate_param, page),
+            "@type": "PartialCollectionView",
+            "first": "{}{}=1".format(iri, paginate_param),
+            "last": "{}{}={}".format(iri, paginate_param, last)
+        }
+        if page != 1:
+            collection_template["view"]["previous"] = "{}{}={}".format(iri, paginate_param,
+                                                                       page-1)
+        if page != last:
+            collection_template["view"]["next"] = "{}{}={}".format(iri, paginate_param,
+                                                                   page + 1)
